Slice 5+7a: the game server's DB thread takes a DatabaseConnection
Every function in `db.rs` now takes `&DatabaseConnection`; the ones still on
hand-written SQL borrow the same pool back out of it
(`get_sqlite_connection_pool`). One pool, two views — the DB thread never opens
a second connection to the file, and the port can proceed function by function
instead of in one 200-query jump.

First loaders ported to entities: `account_gsdata`, `account_premium`,
`lottery` (round + draws), `mdt_history`, `mdt_bets`. `verify_schema` too — it
still fingerprints the database by looking for `characters` and `accounts`.

`lottery`'s prize columns and `mdt_bets.bet` are declared 32-bit in the DDL but
read as `i64` by the game code, so they join the entity type overrides in
`normalize_entities.py`: SQLite stores whatever fits regardless of the declared
type, and a prize past 2^31 would decode-fail against an `i32` field.

2061 game-server tests pass.

+++ crates/models/src/entity/mdt_bets.rs
@@ -7,7 +7,7 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub lane_id: i32,
-    pub bet: Option<i32>,
+    pub bet: Option<i64>,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
